no negatives

+++ include/serialize_data.h
@@ -13,8 +13,8 @@ cJSON* serialize_land_state_to_json(LandState* land_state);
 void deserialize_land_state_from_json(LandState* land_state, cJSON* json);
 cJSON* serialize_units_sea_to_json(UnitsSea* units_sea);
 void deserialize_units_sea_from_json(UnitsSea* units_sea, cJSON* json);
-void set_land_state_field(cJSON* json, const char* key, uint8_t* field);
+void set_land_state_field(cJSON* json, const char* key, uint* field);
 void set_land_state_signed_field(cJSON* json, const char* key, int8_t* field);
-void extract_and_assign(cJSON* json, const char* key, uint8_t* target_array);
+void extract_and_assign(cJSON* json, const char* key, uint* target_array);
 
 #endif
